+++ MacPacker/Features/ArchiveContentViewer/ArchiveContentToolbarView.swift
@@ -34,7 +34,11 @@ struct ArchiveContentToolbarView: ToolbarContent {
                     appDelegate.openPreviewerWindow(for: url)
                 }
             } label: {
-                Label("Preview", systemImage: "doc.text.magnifyingglass")
+                Label {
+                    Text("Preview", comment: "Button in the tooblar that allows the user to preview the selected file.")
+                } icon: {
+                    Image(systemName: "doc.text.magnifyingglass")
+                }
             }
             
             Button {
@@ -42,7 +46,11 @@ struct ArchiveContentToolbarView: ToolbarContent {
                     isExportingItem.toggle()
                 }
             } label: {
-                Label("Extract selected", image: "custom.document.badge.arrow.down")
+                Label {
+                    Text("Extract selected", comment: "Button in the tooblar that allows the user to extract the selected files.")
+                } icon: {
+                    Image("custom.document.badge.arrow.down")
+                }
             }
             .fileImporter(
                 isPresented: $isExportingItem,
@@ -65,7 +73,11 @@ struct ArchiveContentToolbarView: ToolbarContent {
                     isExportingAll.toggle()
                 }
             } label: {
-                Label("Extract archive", image: "custom.shippingbox.badge.arrow.down")
+                Label {
+                    Text("Extract archive", comment: "Button in the toolbar that allows the user to extract the full archive to a target directory.")
+                } icon: {
+                    Image("custom.shippingbox.badge.arrow.down")
+                }
             }
             .fileImporter(
                 isPresented: $isExportingAll,
@@ -82,18 +94,26 @@ struct ArchiveContentToolbarView: ToolbarContent {
                 }
             }
         
-            Menu("More", systemImage: "ellipsis.circle") {
+            Menu {
                 if #available(macOS 14, *) {
                     SettingsLink() {
-                        Label("Settings...", systemImage: "gear")
-                            .labelStyle(.titleAndIcon)
+                        Label {
+                            Text("Settings...", comment: "Used to open the settings/preferences window")
+                        } icon: {
+                            Image(systemName: "gear")
+                        }
+                        .labelStyle(.titleAndIcon)
                     }
                 } else {
                     Button {
                         NSApp.sendAction(Selector(("showSettingsWindow:")), to: nil, from: nil)
                     } label: {
-                        Label("Settings", systemImage: "gear")
-                            .labelStyle(.titleAndIcon)
+                        Label {
+                            Text("Settings...", comment: "Used to open the settings/preferences window")
+                        } icon: {
+                            Image(systemName: "gear")
+                        }
+                        .labelStyle(.titleAndIcon)
                     }
                 }
                 
@@ -104,8 +124,12 @@ struct ArchiveContentToolbarView: ToolbarContent {
                         contentService.openGetInfoWnd(for: [url])
                     }
                 } label: {
-                    Label("Archive info", systemImage: "info.circle")
-                        .labelStyle(.titleAndIcon)
+                    Label {
+                        Text("Archive info", comment: "Used to open Quick Look feature for the current archive file")
+                    } icon: {
+                        Image(systemName: "info.circle")
+                    }
+                    .labelStyle(.titleAndIcon)
                 }
                 
                 Divider()
@@ -114,40 +138,62 @@ struct ArchiveContentToolbarView: ToolbarContent {
                     Button {
                         openURL(URL(string: "https://github.com/sarensw/MacPacker/issues/new?assignees=&labels=enhancement&projects=&template=&title=")!)
                     } label: {
-                        Label("... request a Feature", systemImage: "shippingbox")
-                            .labelStyle(.titleAndIcon)
+                        Label {
+                            Text("... request a Feature", comment: "This is the second part of the text 'Go here to ...'. It is used in the archive window 'More' menu and shall give users a hint about the secondary options to reach out to the dev.")
+                        } icon: {
+                            Image(systemName: "shippingbox")
+                        }
+                        .labelStyle(.titleAndIcon)
                     }
                     
                     Button {
                         openURL(URL(string: "https://github.com/sarensw/MacPacker/issues/new?assignees=&labels=bug&projects=&template=bug_report.md&title=")!)
                     } label: {
-                        Label("... raise a Bug", systemImage: "ladybug")
-                            .labelStyle(.titleAndIcon)
+                        Label {
+                            Text("... raise a Bug", comment: "This is the second part of the text 'Go here to ...'. It is used in the archive window 'More' menu and shall give users a hint about the secondary options to reach out to the dev.")
+                        } icon: {
+                            Image(systemName: "ladybug")
+                        }
+                        .labelStyle(.titleAndIcon)
                     }
                     
                     Button {
-                        openURL(URL(string: "mailto:apps@sarensw.com")!)
+                        openURL(URL(string: "mailto:\(Constants.supportMail)")!)
                     } label: {
-                        Label("... send a mail to apps@sarensw.com", systemImage: "mail")
-                            .labelStyle(.titleAndIcon)
+                        Label {
+                            Text("... send a mail to \(Constants.supportMail)", comment: "This is the second part of the text 'Go here to ...'. It is used in the archive window 'More' menu and shall give users a hint about the secondary options to reach out to the dev.")
+                        } icon: {
+                            Image(systemName: "mail")
+                        }
+                        .labelStyle(.titleAndIcon)
                     }
                 } label: {
-                    Label("Go here to...", systemImage: "exclamationmark.bubble")
-                        .labelStyle(.titleAndIcon)
+                    Label {
+                        Text("Go here to ...", comment: "This is the menu in the 'More' menu of the archive window to give customers a hint what they can do to reach the dev. A submenu will open with links to GitHub, a bug report form, and a mail to the developer.")
+                    } icon: {
+                        Image(systemName: "exclamationmark.bubble")
+                    }
+                    .labelStyle(.titleAndIcon)
                 }
                 
                 Divider()
                 
-                Menu("More Apps", systemImage: "plus.square.dashed") {
+                Menu {
                     Button {
                         openURL(URL(string: "https://filefillet.com/?utm_source=macpacker&utm_content=moremenu&utm_medium=ui")!)
                     } label: {
                         Label {
-                            Text("FileFillet")
+                            Text(Constants.otherAppFileFillet)
                         } icon: {
                             Image(nsImage: .menuIcon(named: "FileFillet", pointSize: 16))
                         }
                         .labelStyle(.titleAndIcon)
+                    }
+                } label: {
+                    Label {
+                        Text("More Apps", comment: "Hint to the user that the submenu contains links for more apps that they might like.")
+                    } icon: {
+                        Image(systemName: "plus.square.dashed")
                     }
                 }
                 .labelStyle(.titleAndIcon)
@@ -155,22 +201,36 @@ struct ArchiveContentToolbarView: ToolbarContent {
                 Button {
                     openURL(URL(string: "https://macpacker.app/?utm_source=macpacker&utm_content=moremenu&utm_medium=ui")!)
                 } label: {
-                    Label("Website", systemImage: "link")
+                    Label {
+                        Text("Website", comment: "Hint to the user that the button links to the app's website.")
+                    } icon: {
+                        Image(systemName: "link")
+                    }
                         .labelStyle(.titleAndIcon)
                 }
                 
                 Button {
                     openURL(URL(string: "https://github.com/sarensw/MacPacker/")!)
                 } label: {
-                    Label("GitHub", systemImage: "link")
+                    Label(Constants.otherAppGitHub, systemImage: "link")
                         .labelStyle(.titleAndIcon)
                 }
                 
                 Button {
                     appDelegate.openAboutWindow()
                 } label: {
-                    Label("About MacPacker", systemImage: "info.square")
-                        .labelStyle(.titleAndIcon)
+                    Label {
+                        Text("About \(Bundle.main.appName)")
+                    } icon: {
+                        Image(systemName: "info.square")
+                    }
+                    .labelStyle(.titleAndIcon)
+                }
+            } label: {
+                Label {
+                    Text("More", comment: "The 'More' menu in the archive window")
+                } icon: {
+                    Image(systemName: "ellipsis")
                 }
             }