+++ MacPacker/Views/PreferencesView.swift
@@ -15,16 +15,24 @@ struct PreferencesView: View {
                 GeneralSettingsView()
                     .tabItem {
                         Image(systemName: "gear")
-                        Text("General")
+                        Text("General", comment: "General settings")
                     }
                     .tag(0)
                 
                 AdvancedSettingsView()
                     .tabItem {
                         Image(systemName: "exclamationmark.octagon")
-                        Text("Advanced")
+                        Text("Advanced", comment: "Advanced settings")
                     }
                     .tag(1)
+                
+                #if DEBUG
+                DebugSettingsView()
+                    .tabItem {
+                        Image(systemName: "ant")
+                        Text(verbatim: "Debug")
+                    }
+                #endif
             }
         }
         .frame(width: 640)