+++ MacPacker/Views/WelcomeView.swift
@@ -8,10 +8,59 @@
 import Foundation
 import SwiftUI
 
+struct WhatsNewPill: View {
+    enum PillType {
+        case feature
+        case bug
+        case core
+    }
+    
+    var title: LocalizedStringResource
+    var type: PillType
+    
+    var body: some View {
+        HStack(alignment: .center, spacing: 8) {
+            switch type {
+            case .feature:
+                Capsule()
+                    .fill(.green.opacity(0.4))
+                    .stroke(.green, lineWidth: 1)
+                    .frame(width: 32, height: 16)
+                    .overlay {
+                        Text(verbatim: "feat")
+                            .font(.footnote)
+                    }
+            case .bug:
+                Capsule()
+                    .fill(.red.opacity(0.4))
+                    .stroke(.red, lineWidth: 1)
+                    .frame(width: 32, height: 16)
+                    .overlay {
+                        Text(verbatim: "bug")
+                            .font(.footnote)
+                    }
+            case .core:
+                Capsule()
+                    .fill(.blue.opacity(0.4))
+                    .stroke(.blue, lineWidth: 1)
+                    .frame(width: 32, height: 16)
+                    .overlay {
+                        Text(verbatim: "core")
+                            .font(.footnote)
+                    }
+            }
+            
+            Text(title)
+        }
+        
+    }
+}
+
 struct WelcomeWhatsNewView: View {
     var body: some View {
         VStack(alignment: .leading, spacing: 4) {
-            Text("• feat: Finder integration via context menu")
+            WhatsNewPill(title: "Finder integration via context menu", type: .feature)
+            WhatsNewPill(title: "Language support for German, French, Simplified Chinese", type: .core)
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
         .padding()
@@ -26,9 +75,9 @@ struct WelcomeOtherProjects: View {
                     .resizable()
                     .frame(width: 32, height: 32)
                 Link(destination: URL(string: "https://filefillet.com/?utm_source=macpacker&utm_content=welcome&utm_medium=ui")!) {
-                    Text("FileFillet")
+                    Text(Constants.otherAppFileFillet)
                 }
-                Text("Organize files without tons of Finder windows.")
+                Text("Organize files without tons of Finder windows.", comment: "Short description of the app FileFillet which is referenced in the 'Other projects' section of the 'Welcome' view")
             }
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
@@ -45,17 +94,17 @@ struct WelcomeFeedbackView: View {
                 ZStack {
                     if feedbackText.isEmpty {
                         VStack {
-                            Text("Please support the archive type ...")
-                            Text("I found the following bug")
-                            Text("Please add feature ...")
+                            Text("Please support the archive type ...", comment: "Hint on what to ask for in the feedback form")
+                            Text("I found the following bug ...", comment: "Hint on what to ask for in the feedback form")
+                            Text("Please add feature ...", comment: "Hint on what to ask for in the feedback form")
                         }
                     }
                     TextEditor(text: $feedbackText)
                         .frame(maxHeight: .infinity)
                         .opacity(feedbackText.isEmpty ? 0.25 : 1)
                 }
-                Button("Send") {
-                    var urlc = URLComponents(string: "mailto:apps@sarensw.com")
+                Button {
+                    var urlc = URLComponents(string: "mailto:\(Constants.supportMail)")
                     urlc?.queryItems = [
                         URLQueryItem(name: "subject", value: "MacPacker Feedback \(UUID())"),
                         URLQueryItem(name: "body", value: feedbackText)
@@ -63,14 +112,20 @@ struct WelcomeFeedbackView: View {
                     if let url = urlc?.url {
                         NSWorkspace.shared.open(url)
                     }
+                } label: {
+                    Text("Send", comment: "Button label to send feedback")
                 }
             }
             
             Spacer()
             HStack(spacing: 14) {
-                Text("Or reach out via...")
-                Link("apps@sarensw.com", destination: URL(string: "mailto:apps@sarensw.com")!)
-                Link("@sarensw", destination: URL(string: "https://twitter.com/sarensw")!)
+                Text("Or reach out via...", comment: "Hint on how to contact the developer")
+                Link(destination: URL(string: "mailto:\(Constants.supportMail)")!) {
+                    Text(verbatim: "\(Constants.supportMail)")
+                }
+                Link(destination: URL(string: "https://twitter.com/sarensw")!) {
+                    Text(verbatim: "@sarensw")
+                }
             }
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
@@ -81,7 +136,22 @@ struct WelcomeFeedbackView: View {
 struct WelcomeView: View {
     @Environment(\.openURL) private var openURL
     @State private var defaultTab = 1
-
+    
+    var styledString: AttributedString {
+        var string = AttributedString(
+            localized: "Welcome to \(Bundle.main.appName)",
+            comment: "'Welcome to <app name>'. The order of the greeting might be different in different languages. For example: English: Welcome to MacPacker, Japanese: MacPackerへようこそ. This is up to the translator to decide."
+        )
+        string.foregroundColor = .secondary
+        string.font = .system(size: 42, weight: .medium)
+        
+        if let range = string.range(of: Bundle.main.appName) {
+            string[range].foregroundColor = .primary
+        }
+        
+        return string
+    }
+    
     var body: some View {
         VStack (alignment: .center, spacing: 8) {
             Group {
@@ -89,12 +159,8 @@ struct WelcomeView: View {
                     .resizable()
                     .aspectRatio(contentMode: .fit)
                     .frame(width: 80, alignment: .center)
-                Text("Welcome to")
-                    .font(.system(size: 42, weight: .medium, design: .default))
-                    .foregroundColor(.secondary) +
-                Text(" MacPacker")
-                    .font(.system(size: 42, weight: .medium, design: .default))
-                Text("Version v\(Bundle.main.appVersionLong)")
+                Text(styledString)
+                Text("Version v\(Bundle.main.appVersionLong)", comment: "This text shows the current version of the app in Welcome and About window")
                     .foregroundColor(.secondary)
             }
             
@@ -102,17 +168,17 @@ struct WelcomeView: View {
                 TabView(selection: $defaultTab) {
                     WelcomeWhatsNewView()
                         .tabItem {
-                            Text("What's new")
+                            Text("What's new", comment: "Title of the tab in the welcome screen that shows the user what's new in this version.")
                         }
                         .tag(1)
                     WelcomeOtherProjects()
                         .tabItem {
-                            Text("More apps")
+                            Text("More apps", comment: "Title of the tab in the welcome screen that links to other projects.")
                         }
                         .tag(2)
                     WelcomeFeedbackView()
                         .tabItem {
-                            Text("Feedback")
+                            Text("Feedback", comment: "Title of the tab in the welcome screen that lets the user give feedback")
                         }
                         .tag(3)
                 }
@@ -122,17 +188,17 @@ struct WelcomeView: View {
                 TabView(selection: $defaultTab) {
                     WelcomeWhatsNewView()
                         .tabItem {
-                            Text("What's new")
+                            Text("What's new", comment: "Title of the tab in the welcome screen that shows the user what's new in this version.")
                         }
                         .tag(1)
                     WelcomeOtherProjects()
                         .tabItem {
-                            Text("More apps")
+                            Text("More apps", comment: "Title of the tab in the welcome screen that links to other projects.")
                         }
                         .tag(2)
                     WelcomeFeedbackView()
                         .tabItem {
-                            Text("Feedback")
+                            Text("Feedback", comment: "Title of the tab in the welcome screen that lets the user give feedback")
                         }
                         .tag(3)
                 }
@@ -141,7 +207,7 @@ struct WelcomeView: View {
             }
             
             #if !STORE
-            Text("Support this Open Source project...")
+            Text("Support the development...")
                 .fontWeight(.semibold)
                 .padding(.top, 14)
             HStack {